pyproject.toml
[tool.poetry]
name = "mychild"
version = "0.1.0"
description = ""
authors = ["Samat Yunusov <samat.yunusov@ccsteam.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.2"
uvicorn = "^0.11.5"
fastapi = "^0.54.1"
databases = "^0.3.2"
tortoise-orm = "^0.16.10"
asyncpg = "^0.20.1"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
